backend/requirements.txt
# Backend API and core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
email-validator>=2.0.0

# Database
psycopg2-binary>=2.9.0

# Auth
bcrypt>=4.0.0
PyJWT>=2.0.0

# Agents (jobs_agent) and store
google-adk[extensions]>=0.1.0
chromadb>=0.4.0
boto3>=1.34.0
beautifulsoup4>=4.12.0

# HTTP (used by tools + tests)
httpx>=0.27.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
